Crate jamendo

jamendo-rs is a client for v3 of the Jamendo API. It exposes a client, whose methods return an action that can be used to specify query parameters.

let client = jamendo::Client::new(jamendo::TEST_ID);
let albums = client.get_albums().limit(15).featured(true).unwrap();

for album in albums {
  println!("{}", album.name);
}

To create a new client, you need a client ID. The test ID changes regularly, for productive usage you should get your own at https://developer.jamendo.com/v3.0/authentication.

Structs

Action
Action represents the request to a resource

Album
A album resource

Artist
A artist resource

Client
The Client is used to create actions

Track
A track resource

UsersAlbums

UsersArtists

UsersTracks

Enums

Error

Query
A query parameter

Resource
A resource endpoint

Constants

TEST_ID
The test client id. Changes regularly, do not use it in a real application!